+++ 6-period/java-oncall-6-main/src/main/java/oncall/controller/OnCallController.java
@@ -4,8 +4,8 @@ import java.util.List;
 import oncall.domain.date.DayType;
 import oncall.domain.date.Month;
 import oncall.domain.date.MonthType;
-import oncall.domain.name.Turn;
-import oncall.domain.name.Turns;
+import oncall.domain.turn.Turn;
+import oncall.domain.turn.Turns;
 import oncall.dto.TurnDto;
 import oncall.exception.ErrorMessage;
 import oncall.exception.ExceptionHandler;
