+++ 6-period/java-oncall-6-main/src/main/java/oncall/domain/turn/Turns.java
@@ -1,4 +1,4 @@
-package oncall.domain.name;
+package oncall.domain.turn;
 
 import oncall.exception.CustomIllegalArgumentException;
 import oncall.exception.ErrorMessage;
